+++ jsolex/src/main/java/me/champeau/a4j/jsolex/app/jfx/bass2000/Bass2000SubmissionController.java
@@ -712,9 +712,9 @@ public class Bass2000SubmissionController {
         var ref = reference != null ? String.format(Locale.US, ", ref:%.3f", reference) : "";
         var output = executor.execute(String.format(Locale.US, """
             [outputs]
-            __result=autocrop2(img(-a2px(a: %.3f%s));1.2)
+            geometryCorrected=autocrop2(img(-a2px(a: %.3f%s));1.2)
             """, shift, ref), ImageMathScriptExecutor.SectionKind.SINGLE);
-        var result = output.imagesByLabel().get("__result");
+        var result = output.imagesByLabel().get("geometryCorrected");
 
         if (result instanceof ImageWrapper image) {
             return image.unwrapToMemory();